refactor(UpcomingEvents): extract types, constants, and date utils into dedicated files
