+++ apps/website/components/UpcomingEvents/constants.ts
@@ -1,0 +1,17 @@
+export const EVENT_TYPE_COLORS: Record<string, string> = {
+  WEEKLY_CALL: 'border-blue-500/30 bg-blue-500/10 text-blue-400',
+  PROJECT_REVIEW: 'border-violet-500/30 bg-violet-500/10 text-violet-400',
+  MENTORSHIP_SESSION: 'border-emerald-500/30 bg-emerald-500/10 text-emerald-400',
+  MENTOR_1V1: 'border-cyan-500/30 bg-cyan-500/10 text-cyan-400',
+  COMMUNITY_MEETING: 'border-amber-500/30 bg-amber-500/10 text-amber-400',
+  SPECIAL_EVENT: 'border-rose-500/30 bg-rose-500/10 text-rose-400',
+};
+
+export const EVENT_TYPE_LABELS: Record<string, string> = {
+  WEEKLY_CALL: 'Weekly Call',
+  PROJECT_REVIEW: 'Project Review',
+  MENTORSHIP_SESSION: 'Mentorship',
+  MENTOR_1V1: '1v1 Session',
+  COMMUNITY_MEETING: 'Community',
+  SPECIAL_EVENT: 'Special Event',
+};
