+++ apps/website/components/UpcomingEvents/utils.ts
@@ -1,0 +1,20 @@
+export function formatEventDate(dateStr: string) {
+  const date = new Date(dateStr);
+  return date.toLocaleDateString('en-US', {
+    weekday: 'short',
+    month: 'short',
+    day: 'numeric',
+  });
+}
+
+export function formatEventTime(startStr: string, endStr: string) {
+  const start = new Date(startStr);
+  const end = new Date(endStr);
+  return `${start.toLocaleTimeString('en-US', { hour: 'numeric', minute: '2-digit' })} - ${end.toLocaleTimeString('en-US', { hour: 'numeric', minute: '2-digit' })}`;
+}
+
+export function isEventToday(dateStr: string) {
+  const eventDate = new Date(dateStr);
+  const now = new Date();
+  return eventDate.toDateString() === now.toDateString();
+}
